rename buffer to bufferOperation
