+++ denops/aider/main.ts
@@ -1,8 +1,8 @@
 import * as fn from "https://deno.land/x/denops_std@v6.4.0/function/mod.ts";
 import type { Denops } from "https://deno.land/x/denops_std@v6.4.0/mod.ts";
 import * as aiderCommand from "./aiderCommand.ts";
-import * as buffer from "./buffer.ts";
-import type { BufferLayout } from "./buffer.ts";
+import * as buffer from "./bufferOperation.ts";
+import type { BufferLayout } from "./bufferOperation.ts";
 import { getCurrentFilePath } from "./utils.ts";
 
 /**
