+++ tests/assertions.ts
@@ -1,6 +1,6 @@
 import { assert } from "https://deno.land/std@0.217.0/assert/assert.ts";
 import type { Denops } from "https://deno.land/x/denops_std@v6.4.0/mod.ts";
-import * as buffer from "../denops/aider/buffer.ts";
+import * as buffer from "../denops/aider/bufferOperation.ts";
 import * as aiderCommand from "../denops/aider/aiderCommand.ts";
 
 export const sleep = (msec: number) =>
